clickhouse_analytics_agent/requirements.txt
# ─── LangGraph & LangChain ────────────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0          # ChatOpenAI adapter (used for OpenRouter)

# ─── ClickHouse ───────────────────────────────────────────────────────────────
clickhouse-connect>=0.7.0         # Official Python ClickHouse HTTP driver

# ─── Data processing ──────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0                   # Parquet read/write (preserves ClickHouse types)
tabulate>=0.9.0                   # df.to_markdown() support

# ─── Visualization ────────────────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ─── API server ───────────────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6           # Required for FastAPI form data

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
